// src/types/api.ts
export interface components {
  schemas: {
    Task: {
      /** @example 1 */
      id?: number;
      /** @example Сделать домашку по запросам */
      name?: string;
      /** @example Использовать патерн controller */
      info?: string;
      /** @example true */
      isImportant?: boolean;
      /** @example false */
      isCompleted?: boolean;
    };
    CreateTask: {
      /** @example Новая задача */
      name: string;
      /** @example Описание задачи */
      info?: string;
      /** @example false */
      isImportant?: boolean;
      /** @example false */
      isCompleted?: boolean;
    };
    UpdateTask: {
      /** @example Обновленное название */
      name?: string;
      /** @example Обновленное описание */
      info?: string;
      /** @example true */
      isImportant?: boolean;
      /** @example true */
      isCompleted?: boolean;
    };
    Error: {
      /** @example Сообщение об ошибке */
      error?: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}